+++ provider-utils/awscloudformation/schemas/vod.schema.graphql.ejs
