+++ provider-utils/awscloudformation/service-walkthroughs/vod-push.js
@@ -287,27 +287,7 @@ async function serviceQuestions(context, options, defaultValuesFilename, resourc
   };
 
   if (cmsResponse.enableCMS) {
-    let apiName = getAPIName(context);
-    if (apiName === '') {
-      await setupAPI(context, options, 'vod' );
-      apiName = getAPIName(context);
-    } else {
-      context.print.info(`Using ${apiName} to manage API`);
-    }
-
-    //await createCMS(context, apiName, props);
-    props.parameters.GraphQLAPIId = {
-      'Fn::GetAtt': [
-        `api${apiName}`,
-        'Outputs.GraphQLAPIIdOutput',
-      ],
-    };
-    props.parameters.GraphQLEndpoint = {
-      'Fn::GetAtt': [
-        `api${apiName}`,
-        'Outputs.GraphQLAPIEndpointOutput',
-      ],
-    };
+    props = await setupAPI(context, props, 'vod');
   }
 
   return props;
@@ -406,233 +386,4 @@ async function createCDNEnvVars(context, options, resourceName, aws) {
   cdnEnvConfigDetails.secretPem = secretCreate.Name;
   cdnEnvConfigDetails.secretPemArn = secretCreate.ARN;
   amplify.saveEnvResourceParameters(context, 'video', resourceName, cdnEnvConfigDetails);
-}
-
-async function createCMS(context, apiName, props) {
-  const permissions = [
-    {
-      type: question.permissionSchema.type,
-      name: question.permissionSchema.key,
-      message: question.permissionSchema.question,
-      choices: question.permissionSchema.options,
-      validate(answer) {
-        if (answer.length < 1) {
-          return 'You must choose at least one auth style';
-        }
-        return true;
-      },
-    },
-  ];
-  const cmsEdit = [
-    {
-      type: question.overrideSchema.type,
-      name: question.overrideSchema.key,
-      message: question.overrideSchema.question,
-      default: question.overrideSchema.default,
-    },
-    {
-      type: question.editAPI.type,
-      name: question.editAPI.key,
-      message: question.editAPI.question,
-      default: question.editAPI.default,
-    }];
-  const backEndDir = context.amplify.pathManager.getBackendDirPath();
-  const resourceDir = path.normalize(path.join(backEndDir, 'api', apiName));
-  let authConfig = {};
-  const amplifyMeta = context.amplify.getProjectMeta();
-  if ('api' in amplifyMeta && Object.keys(amplifyMeta.api).length !== 0) {
-    Object.values(amplifyMeta.api).forEach((project) => {
-      if ('output' in project) {
-        ({ authConfig } = project.output);
-      }
-    });
-  }
-  const permissionsResponse = await inquirer.prompt(permissions);
-  props.permissions = permissionsResponse;
-
-  if (fs.existsSync(`${resourceDir}/schema.graphql`)) {
-    const currentSchema = fs.readFileSync(`${resourceDir}/schema.graphql`);
-    if (!currentSchema.includes('videoObject') && !currentSchema.includes('vodAsset')) {
-      const parameters = JSON.parse(fs.readFileSync(`${resourceDir}/parameters.json`));
-      const cmsEditResponse = await inquirer.prompt(cmsEdit);
-      const editSchemaChoice = cmsEditResponse.editAPI;
-
-      props.cms = cmsEditResponse;
-
-      await writeNewModel(resourceDir, props);
-
-      if (editSchemaChoice) {
-        await context.amplify.openEditor(context, `${resourceDir}/schema.graphql`).then(async () => {
-          let notCompiled = true;
-          while (notCompiled) {
-            notCompiled = await compileSchema(context, resourceDir, parameters, authConfig);
-          }
-        });
-      } else {
-        await compileSchema(context, resourceDir, parameters, authConfig);
-      }
-    } else {
-      const fullPath = path.join(resourceDir, 'schema.graphql');
-      context.print.warning(`Schema already configure. To edit it please open: ${fullPath}`);
-    }
-    // TODO: Add check if they switched schemas
-  }
-  if (props.permissions.permissionSchema.includes('admin')) {
-    authGroupHack(context, props.shared.bucketInput);
-  }
-  createDependency(context, props, apiName);
-}
-
-async function writeNewModel(resourceDir, props) {
-  const appendSchemaTemplate = await fs.readFileSync(`${__dirname}/../schemas/schema.graphql.ejs`, { encoding: 'utf-8' });
-  const appendSchema = ejs.render(appendSchemaTemplate, props);
-
-  if (props.cms.overrideSchema) {
-    await fs.writeFileSync(`${resourceDir}/schema.graphql`, appendSchema);
-  } else {
-    await fs.appendFileSync(`${resourceDir}/schema.graphql`, appendSchema);
-  }
-}
-
-async function createDependency(context, props, apiName) {
-  if (props.contentDeliveryNetwork.pemID && props.contentDeliveryNetwork.secretPemArn) {
-    context.amplify.updateamplifyMetaAfterResourceUpdate('api', apiName, 'dependsOn', [
-      {
-        category: 'video',
-        resourceName: props.shared.resourceName,
-        attributes: [],
-      },
-    ]);
-  }
-}
-
-async function compileSchema(context, resourceDir, parameters, authConfig) {
-  try {
-    await context.amplify.executeProviderUtils(
-      context,
-      'awscloudformation',
-      'compileSchema',
-      { resourceDir, parameters, authConfig },
-    );
-    return false;
-  } catch (e) {
-    context.print.error('Failed compiling GraphQL schema:');
-    context.print.info(e.message);
-    const continueQuestion = {
-      type: 'input',
-      name: 'pressKey',
-      message: `Correct the errors in schema.graphql and press Enter to re-compile.\n\nPath to schema.graphql:\n${resourceDir}/schema.graphql`,
-    };
-    await inquirer.prompt(continueQuestion);
-    return true;
-  }
-}
-
-function getAPIName(context) {
-  const { amplifyMeta } = context.amplify.getProjectDetails();
-  let apiName = '';
-
-  if (amplifyMeta.api) {
-    const categoryResources = amplifyMeta.api;
-    Object.keys(categoryResources).forEach((resource) => {
-      if (categoryResources[resource].service === 'AppSync') {
-        apiName = resource;
-      }
-    });
-  }
-  return apiName;
-}
-
-async function authGroupHack(context, bucketName) {
-  const userPoolGroupFile = path.join(
-    context.amplify.pathManager.getBackendDirPath(),
-    'auth',
-    'userPoolGroups',
-    'user-pool-group-precedence.json',
-  );
-
-  const amplifyMeta = context.amplify.getProjectMeta();
-
-  if (!('auth' in amplifyMeta) || Object.keys(amplifyMeta.auth).length === 0) {
-    context.print.error('You have no auth projects. Moving on.');
-    return;
-  }
-
-  let resourceName = '';
-
-  Object.keys(amplifyMeta.auth).forEach((authCategory) => {
-    if (amplifyMeta.auth[authCategory].service === 'Cognito') {
-      resourceName = authCategory;
-    }
-  });
-
-  if (fs.existsSync(userPoolGroupFile)) {
-    const userPoolGroup = JSON.parse(fs.readFileSync(userPoolGroupFile));
-    if (userPoolGroup.length === 0) {
-      userPoolGroup.push(generateIAMAdmin(resourceName, bucketName));
-    } else {
-      userPoolGroup.forEach((userGroup, index) => {
-        if (userGroup.groupName === 'Admin') {
-          if (!('customPolicies' in userGroup)) {
-            userGroup.customPolicies = [];
-          }
-
-          const policy = generateIAMAdminPolicy(resourceName, bucketName);
-          if (!userGroup.customPolicies.some(
-            existingPolicy => existingPolicy.PolicyName === policy.PolicyName,
-          )) {
-            userGroup.customPolicies.push(policy);
-          }
-          return;
-        }
-        if (userPoolGroup.length === index + 1) {
-          userPoolGroup.push(generateIAMAdmin(resourceName, bucketName));
-        }
-      });
-    }
-    updateUserPoolGroups(context, userPoolGroup);
-  } else {
-    const admin = generateIAMAdmin(resourceName, bucketName);
-    const userPoolGroupList = [admin];
-    updateUserPoolGroups(context, userPoolGroupList);
-    context.amplify.updateamplifyMetaAfterResourceAdd('auth', 'userPoolGroups', {
-      service: 'Cognito-UserPool-Groups',
-      providerPlugin: 'awscloudformation',
-      dependsOn: [
-        {
-          category: 'auth',
-          resourceName,
-          attributes: ['UserPoolId', 'AppClientIDWeb', 'AppClientID', 'IdentityPoolId'],
-        },
-      ],
-    });
-  }
-}
-
-
-function updateUserPoolGroups(context, userPoolGroupList) {
-  if (userPoolGroupList && userPoolGroupList.length > 0) {
-    const userPoolGroupFile = path.join(
-      context.amplify.pathManager.getBackendDirPath(),
-      'auth',
-      'userPoolGroups',
-      'user-pool-group-precedence.json',
-    );
-
-    const userPoolGroupParams = path.join(context.amplify.pathManager.getBackendDirPath(), 'auth', 'userPoolGroups', 'parameters.json');
-
-    /* eslint-disable */
-    const groupParams = {
-      AuthRoleArn: {
-        'Fn::GetAtt': ['AuthRole', 'Arn'],
-      },
-      UnauthRoleArn: {
-        'Fn::GetAtt': ['UnauthRole', 'Arn'],
-      },
-    };
-    /* eslint-enable */
-
-    fs.outputFileSync(userPoolGroupParams, JSON.stringify(groupParams, null, 4));
-    fs.outputFileSync(userPoolGroupFile, JSON.stringify(userPoolGroupList, null, 4));
-  }
 }
